Move config to root dir
