+++ ops/services/app_functions/report_stream_batched_publisher/functions/QueueBatchedReportStreamUploader/lib.ts
@@ -6,7 +6,7 @@ import {
   StorageSharedKeyCredential,
 } from "@azure/storage-queue";
 import * as csvStringify from "csv-stringify/lib/sync";
-import { ENV, uploaderVersion } from "./config";
+import { ENV, uploaderVersion } from "../config";
 import fetch, { Headers } from "node-fetch";
 import {
   ReportStreamResponse,
