+++ ops/services/app_functions/report_stream_batched_publisher/functions/config.ts
@@ -26,6 +26,8 @@ export const ENV = (() => {
     REPORT_STREAM_BATCH_MINIMUM: "minimum # of messages to read from the queue",
     REPORT_STREAM_BATCH_MAXIMUM:
       "maximum # of messages to send to ReportStream",
+    SIMPLE_REPORT_CB_URL: "URL to the SimpleReport webhook for ReportStream exceptions",
+    SIMPLE_REPORT_CB_TOKEN: "API token to the SimpleReport webhook for ReportStream exceptions"
   };
   return Object.entries(CONFIG_VALUES).reduce((acc, cur) => {
     const [configName, description] = [
